Explore/Homepage Updated 19/02/24
This File contains the updated explore page with the including the product slider. All that needs to be added to the page is the brand and type of wine. Also I have added a button that is with in the product slider that says add to basket. This needs to be made functional. I am waiting for the basket page to become functional.